Removed a huge amount of superfluous vars. cleaned up some functions and moved
